+++ src/main/java/io/github/astrapi69/entity/deletable/ByDeletion.java
@@ -36,7 +36,7 @@ import lombok.Setter;
 import lombok.ToString;
 import lombok.experimental.SuperBuilder;
 import io.github.astrapi69.data.deletable.IdentifiableByDeletable;
-import io.github.astrapi69.entity.base.SequenceBaseEntity;
+import io.github.astrapi69.entity.identifiable.SequenceBaseEntity;
 
 /**
  * The entity class {@link ByDeletion} acts like an deletion flag that keeps the information for the