+++ src/main/java/io/github/astrapi69/entity/modifiable/ByLastModification.java
@@ -36,7 +36,7 @@ import lombok.Setter;
 import lombok.ToString;
 import lombok.experimental.SuperBuilder;
 import io.github.astrapi69.data.modifiable.IdentifiableByLastModified;
-import io.github.astrapi69.entity.base.SequenceBaseEntity;
+import io.github.astrapi69.entity.identifiable.SequenceBaseEntity;
 
 /**
  * The entity class {@link ByLastModification} is keeping information of the last modification of an